refactor: to `mga.py`
